seperate mcp3208 spi dev
Separated the MCP3208 SPI communication part from the gas module.
