+++ mcp3208dev/mknod.sh
@@ -1,0 +1,6 @@
+#! /bin/sh
+
+MODULE="mcp3208dev0.0" # [name}{master}.{chip}
+MAJOR=$(awk "\$2==\"$MODULE\" {print \$1}" /proc/devices)
+
+mknod /dev/$MODULE c $MAJOR 0
